made sure that the pieces were in correctly and combined with origin
