+++ app/src/router/routes.ts
@@ -6,6 +6,8 @@ const routes: RouteRecordRaw[] = [
     component: () => import('layouts/MainLayout.vue'),
     children: [
       { path: '', component: () => import('pages/IndexPage.vue') },
+      { path: 'teaminfo/:teamName', name: 'TeamInfo', component: () => import('pages/TeamInfo.vue') },
+      { path: 'leagueinfo/:id', name: 'LeagueInfo', component: () => import('pages/LeagueInfo.vue') },
       { path: 'chat', component: () => import('pages/chat/chatPage.vue')},
       { path: 'profile', component: () => import('pages/profile/profilePage.vue')},
       { path: 'schedule', component: () => import('pages/schedule/schedulePage.vue')},
